requirements.txt
crewai
python-dotenv
streamlit
pysqlite3-binary>=0.5.1
fastapi
uvicorn
requests